with-mimetypes middleware.
Middleware to add the proper Content-Type header based on the uri of
the request.  options is a map containing a :mimetype map of extension
to type and a :default mime type.  If :mimetype is not provided, a default
map with common mime types will be used.  If :default is not provided,
"text/html" is used.